requirements.txt

phidata>=2.4.0              # Core framework
openai>=1.6.0               # LLM providers (pick what you use)
tiktoken>=0.5.2             # Embeddings & RAG utilities
groq>=0.9.0
numpy>=1.26.0
sqlalchemy>=2.0.25          # Database / pgvector
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0        # Environment & config
asyncio                     # Async / concurrency (agents)
aiohttp>=3.9.1
loguru>=0.7.2               # Logging & observability
pydantic>=2.5.2             # Utilities
tenacity>=8.2.3
rich>=13.7.0
yfinance>=0.2.28           # Financial data
packaging>=23.15           # Version parsing
duckduckgo-search>=2.6.0      # Search engine integration
fastapi>=0.100.0           # API framework
uvicorn[standard]>=0.23.2   # ASGI server
python-multipart>=0.0.6     # File uploads
grokking==0.0.7             # Text parsing utilities
psycopg[binary]>=2.9.9     # PostgreSQL driver
chromadb>=0.3.21              # Vector database